benchmark: rename arch_bench_routes to bench_routes
The routes module is the prompt-construction registry every pipeline
layer consumes, not an arch-bench-specific engine part -- the experiment
orchestrator this PR adds imports it, so it takes the pipeline naming
convention here. Pure rename; consumers' import lines updated in the
same commit.
