+++ tests/test_bench_dry_run.py
@@ -18,7 +18,7 @@ from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_d
     generate_mock_parsed_output,
     generate_mock_raw_outputs,
 )
-from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.arch_bench_routes import (
+from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_routes import (
     ROUTE_REGISTRY,
 )
