+++ tests/test_wording_bench.py
@@ -21,7 +21,7 @@ _REPO_ROOT = Path(__file__).resolve().parents[1]
 if str(_REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(_REPO_ROOT))
 
-from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.arch_bench_routes import (
+from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.bench_routes import (
     ROUTE_REGISTRY,
 )
 from tests.phase1_prompt_benchmarking.architecture_benchmarking_workflow.wording_bench_alternates import (
